requirements.txt
streamlit>=1.32
openai>=1.0.0
pydantic>=2.6
python-dotenv>=1.0.1
pandas>=2.2
